change MAX_ALIGNMENT to double, as requested by djb@cr.yp.to

+++ gnu/usr.bin/ld/i386/md.h
@@ -1,4 +1,4 @@
-/* *	$OpenBSD: md.h,v 1.3 1998/03/26 19:46:40 niklas Exp $*/
+/* *	$OpenBSD: md.h,v 1.4 2000/02/01 21:01:39 espie Exp $*/
 /*
  * Copyright (c) 1993 Paul Kranenburg
  * All rights reserved.
@@ -39,7 +39,7 @@
 #endif
 #endif
 
-#define	MAX_ALIGNMENT		(sizeof (long))
+#define	MAX_ALIGNMENT		(sizeof (double))
 
 #if defined(__NetBSD__) || defined(__OpenBSD__)
 #define PAGSIZ			__LDPGSZ
